gw업로드/삭제 연동

+++ gm-api-user/src/main/java/com/nomadconnection/dapp/api/service/UserCorporationService.java
@@ -3,6 +3,7 @@ package com.nomadconnection.dapp.api.service;
 import com.nomadconnection.dapp.api.common.Const;
 import com.nomadconnection.dapp.api.dto.UserCorporationDto;
 import com.nomadconnection.dapp.api.exception.BadRequestedException;
+import com.nomadconnection.dapp.api.exception.EmptyResxException;
 import com.nomadconnection.dapp.api.exception.EntityNotFoundException;
 import com.nomadconnection.dapp.api.exception.MismatchedException;
 import com.nomadconnection.dapp.core.domain.*;
@@ -14,6 +15,7 @@ import com.nomadconnection.dapp.core.domain.repository.shinhan.D1000Repository;
 import com.nomadconnection.dapp.core.domain.repository.shinhan.D1100Repository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.io.FilenameUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.ObjectUtils;
@@ -208,13 +210,14 @@ public class UserCorporationService {
      * 주주명부 파일 등록
      *
      * @param idx_user     등록하는 User idx
-     * @param files         파일
+     * @param files        파일
      * @param type         file type
      * @param idx_CardInfo CardIssuanceInfo idx
+     * @param cardCode     카드코드
      * @return 등록 정보
      */
     @Transactional(rollbackFor = Exception.class)
-    public List<UserCorporationDto.StockholderFileRes> uploadStockholderFile(Long idx_user, MultipartFile[] files, String type, Long idx_CardInfo) {
+    public List<UserCorporationDto.StockholderFileRes> uploadStockholderFile(Long idx_user, MultipartFile[] files, String type, Long idx_CardInfo, String cardCode) throws IOException {
         User user = findUser(idx_user);
         CardIssuanceInfo cardInfo = findCardIssuanceInfo(user.corp());
         if (!cardInfo.idx().equals(idx_CardInfo)) {
@@ -225,25 +228,32 @@ public class UserCorporationService {
             throw BadRequestedException.builder().category(BadRequestedException.Category.EXCESS_UPLOAD_FILE_COUNT).build();
         }
 
+        if (files == null || ObjectUtils.isEmpty(files)) {
+            throw EmptyResxException.builder().build();
+        }
+
         if (!ObjectUtils.isEmpty(fileList)) {
             for (StockholderFile file : fileList) {
                 repoFile.delete(file);
+                gwUploadService.delete(file.fname(), cardCode);
                 s3Service.s3FileDelete(file.s3Key());
             }
         }
-
         List<UserCorporationDto.StockholderFileRes> resultList = new ArrayList<>();
         for (MultipartFile file : files) {
-            String fileName = UUID.randomUUID().toString();
+            String fileName = UUID.randomUUID().toString() + "." + FilenameUtils.getExtension(file.getOriginalFilename());
             File uploadFile = new File(fileName);
             try {
                 uploadFile.createNewFile();
                 FileOutputStream fos = new FileOutputStream(uploadFile);
                 fos.write(file.getBytes());
                 fos.close();
-                gwUploadService.upload(uploadFile, "0311");
+
+                gwUploadService.upload(uploadFile, cardCode);
+
                 String s3Key = "stockholder/" + idx_CardInfo + "/" + fileName;
                 String s3Link = s3Service.s3FileUpload(uploadFile, s3Key);
+
                 uploadFile.delete();
 
                 resultList.add(UserCorporationDto.StockholderFileRes.from(repoFile.save(StockholderFile.builder()
@@ -256,9 +266,9 @@ public class UserCorporationService {
                         .size(file.getSize())
                         .orgfname(file.getOriginalFilename()).build()), cardInfo.idx()));
 
-            } catch (IOException e) {
+            } catch (Exception e) {
                 uploadFile.delete();
-                e.printStackTrace();
+                throw e;
             }
         }
